Move PSKReporter class to own component.
